+++ src/main/java/lotto/input/validation/code/BonusNumberValidationCode.java
@@ -1,4 +1,4 @@
-package lotto.validation.code;
+package lotto.input.validation.code;
 
 public enum BonusNumberValidationCode {
     INVALID_FORMAT("입력된 값이 올바른 보너스 숫자 형식이 아닙니다"),
