+++ src/main/java/lotto/input/validation/code/TotalPriceValidationCode.java
@@ -1,4 +1,4 @@
-package lotto.validation.code;
+package lotto.input.validation.code;
 
 public enum TotalPriceValidationCode {
     INVALID_NUMBER_FORMAT("숫자 외 문자를 포함해서는 안 되며, 첫 번째 숫자가 0이 되어서는 안 됩니다."),
